+++ .env.development
@@ -1,0 +1,2 @@
+NOTION_ACCESS_TOKEN=secret_oBWs2VIuNGX0492m8wKnm4gxr100UFRS1IQwkvAIpyf
+NOTION_DATABASE_ID=9771e9b4d34b4f73a56ac5424be7e9b8
