+++ src/components/blocks/NestedList.tsx
@@ -1,4 +1,4 @@
-import RenderBlock from '../../RenderBlock';
+import RenderBlock from './RenderBlock';
 
 export const renderNestedList = (block: any) => {
     const { type } = block;
