+++ src/pages/api/utils/notion.ts
@@ -4,3 +4,5 @@ import { Client } from "@notionhq/client";
 export const notion = new Client({
     auth: process.env.NOTION_ACCESS_TOKEN,
 });
+
+export const databaseId = process.env.NOTION_DATABASE_ID;
